refactor(viz): move phenotypic.viz into phenotypic.tools_.viz
Relocate the shared visualization layer (centralized Plotly theme,
matplotlib rcParams mirror, ipywidgets notebook shell) under tools_ to
sit alongside the other framework utilities. The package was never a
public top-level export and uses only internal relative imports, so the
move is a pure rename plus reference updates.

- git mv src/phenotypic/viz -> src/phenotypic/tools_/viz
- rewrite all phenotypic.viz.* imports to phenotypic.tools_.viz.*
  across abc_, correction, grid, gui, and the plot accessor
- fix path-segment references in tests/unit/viz/test_import_rules.py
- add a viz/ entry to tools_/CLAUDE.md
