+++ tests/unit/measure/test_symmetric_zones_figure.py
@@ -16,7 +16,7 @@ from phenotypic import Image
 from phenotypic.measure import MeasureSymmetricZones
 from phenotypic.measure._measure_symmetric_zones import BASE_LAYER
 from phenotypic.tools_ import CONFIG_SUFFIX_OPERATION, ensure_typed_json_suffix
-from phenotypic.viz.figures._theme import OKABE_ITO
+from phenotypic.tools_.viz.figures._theme import OKABE_ITO
 
 
 def _circular_colony_image() -> Image:
